+++ apps/router-e2e/metro.config.js
@@ -16,8 +16,8 @@ config.watchFolders = [projectRoot, ...['packages', 'node_modules'].map((v) => p
 config.resolver.blockList = [
   ...config.resolver.blockList,
 
-  /node_modules\/@react-navigation\/native-stack\/node_modules\/@react-navigation\//,
   /\/expo-router\/node_modules\/@react-navigation/,
+  /node_modules\/@react-navigation\/native-stack\/node_modules\/@react-navigation\//,
   /node_modules\/pretty-format\/node_modules\/react-is/,
 ];