+++ packages/@expo/cli/src/utils/tsconfig/resolveWithTsConfigPaths.ts
@@ -10,7 +10,7 @@ const debug = require('debug')('expo:metro:tsconfig-paths') as typeof console.lo
 const isAbsolute = process.platform === 'win32' ? path.win32.isAbsolute : path.posix.isAbsolute;
 
 export function resolveWithTsConfigPaths(
-  config: { paths: Paths; baseUrl: string },
+  config: { paths: Paths; baseUrl: string; hasBaseUrl: boolean },
   request: {
     /** Import request */
     moduleName: string;
@@ -23,7 +23,7 @@ export function resolveWithTsConfigPaths(
 
   if (
     // If no aliases are added bail out
-    !aliases.length ||
+    (!aliases.length && !config.hasBaseUrl) ||
     // Library authors cannot utilize this feature in userspace.
     /node_modules/.test(request.originModulePath) ||
     // Absolute paths are not supported
@@ -35,22 +35,32 @@ export function resolveWithTsConfigPaths(
   }
 
   const matched = matchTsConfigPathAlias(aliases, request.moduleName);
-  if (!matched) {
-    return null;
-  }
+  if (matched) {
+    for (const alias of config.paths[matched.text]) {
+      const nextModuleName = matched.star ? alias.replace('*', matched.star) : alias;
 
-  for (const alias of config.paths[matched.text]) {
-    const nextModuleName = matched.star ? alias.replace('*', matched.star) : alias;
+      if (/\.d\.ts$/.test(nextModuleName)) continue;
 
-    if (/\.d\.ts$/.test(nextModuleName)) continue;
+      const possibleResult = path.join(config.baseUrl, nextModuleName);
 
-    const possibleResult = path.join(config.baseUrl, nextModuleName);
-
-    const result = resolve(possibleResult);
-    if (result) {
-      debug(`${request.moduleName} -> ${possibleResult}`);
-      return result;
+      const result = resolve(possibleResult);
+      if (result) {
+        debug(`${request.moduleName} -> ${possibleResult}`);
+        return result;
+      }
+    }
+  } else {
+    // Only resolve against baseUrl if no `paths` groups were matched.
+    // Base URL is resolved after paths, and before node_modules.
+    if (config.hasBaseUrl) {
+      const possibleResult = path.join(config.baseUrl, request.moduleName);
+      const result = resolve(possibleResult);
+      if (result) {
+        debug(`baseUrl: ${request.moduleName} -> ${possibleResult}`);
+        return result;
+      }
     }
   }
+
   return null;
 }